+++ src/main/java/com/amannmalik/mcp/jsonrpc/JsonRpcCodec.java
@@ -92,7 +92,7 @@ public final class JsonRpcCodec {
             throw new IllegalArgumentException("id is required");
         }
         return switch (value.getValueType()) {
-            case NUMBER -> new RequestId.NumericId(((JsonNumber) value).doubleValue());
+            case NUMBER -> new RequestId.NumericId(((JsonNumber) value).longValue());
             case STRING -> new RequestId.StringId(((JsonString) value).getString());
             default -> throw new IllegalArgumentException("Invalid id type");
         };
